+++ Craft.Net.Client/MinecraftClient.Properties.cs
@@ -59,7 +59,7 @@ namespace Craft.Net.Client
 
         #endregion
 
-        public short Health { get; internal set; }
+        public float Health { get; internal set; }
         public short Food { get; internal set; }
         public float FoodSaturation { get; internal set; }
